🚚 Relocate request and response to their directory
