+++ src/main/java/com/und/server/auth/dto/request/RefreshTokenRequest.java
@@ -1,4 +1,4 @@
-package com.und.server.auth.dto;
+package com.und.server.auth.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
