+++ src/main/java/com/und/server/common/dto/request/TestAuthRequest.java
@@ -1,4 +1,4 @@
-package com.und.server.common.dto;
+package com.und.server.common.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
