+++ src/main/java/com/und/server/common/exception/GlobalExceptionHandler.java
@@ -14,7 +14,7 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.context.request.WebRequest;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
 
-import com.und.server.common.dto.ErrorResponse;
+import com.und.server.common.dto.response.ErrorResponse;
 import com.und.server.member.exception.MemberErrorResult;
 
 import io.swagger.v3.oas.annotations.Hidden;
