+++ src/main/java/com/und/server/member/dto/request/NicknameRequest.java
@@ -1,4 +1,4 @@
-package com.und.server.member.dto;
+package com.und.server.member.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
