+++ src/main/java/com/und/server/member/service/MemberService.java
@@ -10,8 +10,8 @@ import com.und.server.auth.exception.AuthErrorResult;
 import com.und.server.auth.oauth.Provider;
 import com.und.server.auth.service.RefreshTokenService;
 import com.und.server.common.exception.ServerException;
-import com.und.server.member.dto.MemberResponse;
-import com.und.server.member.dto.NicknameRequest;
+import com.und.server.member.dto.request.NicknameRequest;
+import com.und.server.member.dto.response.MemberResponse;
 import com.und.server.member.entity.Member;
 import com.und.server.member.exception.MemberErrorResult;
 import com.und.server.member.repository.MemberRepository;
